Fix routes

+++ routes/passport.php
@@ -7,21 +7,10 @@ use Illuminate\Support\Facades\Route;
  *
  * @return void
  */
-    Route::group(['middleware' => ['web', 'portal-auth']], function ($router) {
-        Route::get('/authorize', [
-            'uses' => [\Laravel\Passport\Http\Controllers\AuthorizationController::class, 'authorize'],
-            'as' => 'passport.authorizations.authorize',
-        ]);
-
-        Route::post('/authorize', [
-            'uses' => [\Laravel\Passport\Http\Controllers\ApproveAuthorizationController::class, 'approve'],
-            'as' => 'passport.authorizations.approve',
-        ]);
-
-        Route::delete('/authorize', [
-            'uses' => [\Laravel\Passport\Http\Controllers\DenyAuthorizationController::class, 'deny'],
-            'as' => 'passport.authorizations.deny',
-        ]);
+    Route::group(['middleware' => ['web', 'portal-auth']], function () {
+        Route::get('/authorize', [\Laravel\Passport\Http\Controllers\AuthorizationController::class, 'authorize'])->name('passport.authorizations.authorize');
+        Route::post('/authorize', [\Laravel\Passport\Http\Controllers\ApproveAuthorizationController::class, 'approve'])->name('passport.authorizations.approve');
+        Route::delete('/authorize', [\Laravel\Passport\Http\Controllers\DenyAuthorizationController::class, 'deny'])->name('passport.authorizations.deny');
     });
 
 /**
@@ -29,22 +18,11 @@ use Illuminate\Support\Facades\Route;
  *
  * @return void
  */
-    Route::post('/token', [
-        'uses' => [\Laravel\Passport\Http\Controllers\AccessTokenController::class, 'issueToken'],
-        'as' => 'passport.token',
-        'middleware' => 'throttle',
-    ]);
+    Route::middleware('throttle')->post('/token', [\Laravel\Passport\Http\Controllers\AccessTokenController::class, 'issueToken'])->name('passport.token');
 
-    Route::group(['middleware' => ['web', 'portal-auth']], function ($router) {
-        Route::get('/tokens', [
-            'uses' => [\Laravel\Passport\Http\Controllers\AuthorizedAccessTokenController::class, 'forUser'],
-            'as' => 'passport.tokens.index',
-        ]);
-
-        Route::delete('/tokens/{token_id}', [
-            'uses' => [\Laravel\Passport\Http\Controllers\AuthorizedAccessTokenController::class, 'destroy'],
-            'as' => 'passport.tokens.destroy',
-        ]);
+    Route::group(['middleware' => ['web', 'portal-auth']], function () {
+        Route::get('/tokens', [\Laravel\Passport\Http\Controllers\AuthorizedAccessTokenController::class, 'forUser'])->name('passport.tokens.index');
+        Route::delete('/tokens/{token_id}', [\Laravel\Passport\Http\Controllers\AuthorizedAccessTokenController::class, 'destroy'])->name('passport.tokens.destroy');
     });
 
 /**
@@ -52,64 +30,4 @@ use Illuminate\Support\Facades\Route;
  *
  * @return void
  */
-    Route::post('/token/refresh', [
-        'middleware' => ['web', 'portal-auth'],
-        'uses' => [\Laravel\Passport\Http\Controllers\TransientTokenController::class, 'refresh'],
-        'as' => 'passport.token.refresh',
-    ]);
-
-/**
- * Register the routes needed for managing clients.
- *
- * @return void
- */
-//    Route::group(['middleware' => ['web', 'portal-auth']], function ($router) {
-//        Route::get('/clients', [
-//            'uses' => 'ClientController@forUser',
-//            'as' => 'passport.clients.index',
-//        ]);
-//
-//        Route::post('/clients', [
-//            'uses' => 'ClientController@store',
-//            'as' => 'passport.clients.store',
-//        ]);
-//
-//        Route::put('/clients/{client_id}', [
-//            'uses' => 'ClientController@update',
-//            'as' => 'passport.clients.update',
-//        ]);
-//
-//        Route::delete('/clients/{client_id}', [
-//            'uses' => 'ClientController@destroy',
-//            'as' => 'passport.clients.destroy',
-//        ]);
-//    });
-
-
-
-/**
- * Register the routes needed for managing personal access tokens.
- *
- * @return void
- */
-//    Route::group(['middleware' => ['web', 'portal-auth']], function ($router) {
-//        Route::get('/scopes', [
-//            'uses' => 'ScopeController@all',
-//            'as' => 'passport.scopes.index',
-//        ]);
-//
-//        Route::get('/personal-access-tokens', [
-//            'uses' => 'PersonalAccessTokenController@forUser',
-//            'as' => 'passport.personal.tokens.index',
-//        ]);
-//
-//        Route::post('/personal-access-tokens', [
-//            'uses' => 'PersonalAccessTokenController@store',
-//            'as' => 'passport.personal.tokens.store',
-//        ]);
-//
-//        Route::delete('/personal-access-tokens/{token_id}', [
-//            'uses' => 'PersonalAccessTokenController@destroy',
-//            'as' => 'passport.personal.tokens.destroy',
-//        ]);
-//    });
+    Route::middleware(['web', 'portal-auth'])->post('/token/refresh', [\Laravel\Passport\Http\Controllers\TransientTokenController::class, 'refresh'])->name('passport.token.refresh');
